pyproject.toml
[project]
name = "automate-job-searching-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "docling>=2.67.0",
    "faiss-cpu>=1.13.1",
    "langchain>=1.1.3",
    "langchain-classic>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-ollama>=1.0.1",
    "langchain-text-splitters>=1.1.0",
    "matplotlib>=3.10.8",
    "nltk>=3.9.2",
    "ollama>=0.6.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-jobspy>=1.1.82",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "spacy>=3.8.11",
    "streamlit>=1.52.1",
    "flask-cors>=6.0.2",
    "flask[async]>=3.1.2",
    "langchain-community>=0.4.1",
    "livekit-agents[bey]~=1.3",
    "livekit-api>=1.1.0",
    "livekit-plugins-google>=1.3.6",
    "livekit-plugins-groq>=1.3.6",
    "livekit-plugins-noise-cancellation>=0.2.5",
    "livekit-plugins-silero>=1.3.6",
    "livekit-plugins-turn-detector>=1.3.6",
    "mem0ai>=1.0.1",
    "openai-whisper>=20250625",
    "uvicorn>=0.40.0",
]

[tool.uv]
override-dependencies = [
    "numpy>=2.1.0"
]

